bokeh plots - saving png
