+++ wba-trailcounting/requirements.txt
@@ -9,6 +9,9 @@ flake8
 requests
 beautifulsoup4==4.8.1
 numpy
+selenium
+phantomjs
+pillow
 pandas==0.25.3
 seaborn==0.9.0
 bokeh==1.4.0
